snapgoal-task/requirements.txt
# python 3.10.0

django==4.2.11
python-dotenv==1.0.0
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5
# django-cors-headers==4.3.1
requests==2.31.0
django-cache-memoize==0.2.0
drf-spectacular==0.26.2
